Big update for chicken model.
This update allows the chicken ux to roll in a different way. The
chicken model does not require a search since it is a simple single
class and does not need help on how to remedy since it is not a disease.

The commit entails: pulling the number of detections and passing it to
UX function, creating a seperate UX for chicken. setting a parameter
called model type in settings to understand if it should render seach UX
(weed,crop disease,etc) or a count UX (chicken).